+++ package.json
@@ -13,7 +13,9 @@
   "license": "ISC",
   "dependencies": {
     "express": "^4.19.2",
-    "morgan": "^1.10.0"
+    "express-handlebars": "^7.1.2",
+    "morgan": "^1.10.0",
+    "socket.io": "^4.7.5"
   },
   "devDependencies": {
     "nodemon": "^3.1.0"
