+++ src/data/fs/files/users.json
@@ -11,7 +11,7 @@
         "photo": "./assets/imgpath.jpg",
         "email": "sarahsmith456@email.com",
         "password": "SecurePass789!",
-        "role": 0
+        "role": "0"
     },
     {
         "id": "24e09fe34078c24127a1375c",
@@ -38,7 +38,7 @@
         "id": "fef492c21070e00f464af367",
         "photo": "./assets/imgpath.jpg",
         "email": "davidwilson987@email.com",
-        "password": " WilsonD@v3",
+        "password": "WilsonD@v3",
         "role": "1"
     },
     {
@@ -70,70 +70,70 @@
         "role": "1"
     },
     {
-        "id": "cfc53a46c51c7cb4c6d30a63",
+        "id": "df6e8cf88c05e2b5a0b72b36",
         "photo": "./assets/imgpath.jpg",
         "email": "markthompson123@email.com",
         "password": "M@rkTh0mp!234",
         "role": "0"
     },
     {
-        "id": "f6226aa687336f04cb3b9d8e",
+        "id": "99f0b72b3d9cd3a6cf8c6724",
         "photo": "./assets/imgpath.jpg",
         "email": "jessicablake456@email.com",
         "password": "Bl@keJ3ss!789",
         "role": "1"
     },
     {
-        "id": "24e09fe34078c24127a1375c",
+        "id": "b8e31d3d0fcb9eaff3a9e92d",
         "photo": "./assets/imgpath.jpg",
         "email": "matthewdavis432@email.com",
         "password": "Dav!sM@tth3w1",
         "role": "0"
     },
     {
-        "id": "e9172eca317f8a4b984e307c",
+        "id": "96d9ee0c2ab1e2c91bbdc342",
         "photo": "./assets/imgpath.jpg",
         "email": "ashleyharrison321@email.com",
         "password": "H@rr!s0nAs!ey321",
         "role": "1"
     },
     {
-        "id": "0e2faca1dab7b542ce9c0750",
+        "id": "ab8b3e4c8cf696b6749ed689",
         "photo": "./assets/imgpath.jpg",
         "email": "nathanrobinson789@email.com",
         "password": "R0b!nsonN@th@n789",
         "role": "0"
     },
     {
-        "id": "fef492c21070e00f464af367",
+        "id": "8a06c9e109ecb09d3f2f8457",
         "photo": "./assets/imgpath.jpg",
         "email": "oliviataylor987@email.com",
         "password": "T@yl0r0l!v!a987",
         "role": "1"
     },
     {
-        "id": "6a7ca9f934947bb6952e960b",
+        "id": "152c18bb6ee5da454c02fb1c",
         "photo": "./assets/imgpath.jpg",
         "email": "danielcooper654@email.com",
         "password": "Coop3rD@niel654",
         "role": "0"
     },
     {
-        "id": "09bb4f5edc5b7d5672f7a104",
+        "id": "5b89a3e5ad0dc7c4a4c34351",
         "photo": "./assets/imgpath.jpg",
         "email": "meganturner234@email.com",
         "password": "Turn3rMeg@n234",
         "role": "0"
     },
     {
-        "id": "b38e4a83b15523ca19f7e42c",
+        "id": "8c9d431686bc7ad886363bc2",
         "photo": "./assets/imgpath.jpg",
         "email": "lukepeterson876@email.com",
         "password": "P3t3rs0nLuk3!876",
         "role": "1"
     },
     {
-        "id": "d175674605aaa21d25a82f43",
+        "id": "7c6aa7257d36a0cf60f6ec8d",
         "photo": "./assets/imgpath.jpg",
         "email": "samanthajenkins890@email.com",
         "password": "J3nk!nsS@m@nth@890",
