+++ src/routers/api/products.api.js
@@ -1,25 +1,26 @@
 import { Router } from "express";
-import userManager from '../../src/data/fs/files/UserManager.fs.js'
+import productManager from "../../data/fs/files/ProductManager.fs.js";
 
-const usersRouter = Router();
+const productsRouter = Router();
 
 //routes
-usersRouter.get('/', read);
-usersRouter.get('/:uid', readOne);
-usersRouter.post('/', create);
-usersRouter.put('/:uid', update);
-usersRouter.delete('/:uid', destroy);
+productsRouter.get('/', read);
+productsRouter.get('/:pid', readOne);
+productsRouter.post('/', create);
+productsRouter.put('/:pid', update);
+productsRouter.delete('/:pid', destroy);
+// asigno los distintos métodos con sus respectivos endpoints, y la callback que llaman
 
 //functions
 async function read (req, res, next) {
     try {
-        const { role } = req.query;
-        const allUsers = await userManager.read(role);
-        if (allUsers) {
+        const { category } = req.query;
+        const allProducts = await productManager.read(category);
+        if (allProducts) {
             return res.json({
                 statusCode: 200,
-                response: allUsers,
-                role,
+                response: allProducts,
+                category,
                 success: true
             });
         } else {
@@ -34,8 +35,8 @@ async function read (req, res, next) {
 
 async function readOne (req, res, next) {
     try {
-        const { uid } = req.params;
-        const selected = await userManager.readOne(uid);
+        const { pid } = req.params;
+        const selected = await productManager.readOne(pid);
         if (selected) {
             return res.json({
                 statusCode: 200,
@@ -55,11 +56,11 @@ async function readOne (req, res, next) {
 async function create (req, res, next) {
     try {
         const data = req.body;
-        const newUser = await userManager.create(data)
+        const newProduct = await productManager.create(data)
         return res.json({
             statusCode: 201,
-            message: `User created successfully with id ${newUser.id}`,
-            response: newUser
+            message: `Product created successfully with id ${newProduct.id}`,
+            response: newProduct
         });
     } catch(err) {
         return next(err);
@@ -68,13 +69,13 @@ async function create (req, res, next) {
 
 async function update (req, res, next) {
     try {
-        const { uid } = req.params;
+        const { pid } = req.params;
         const data = req.body;
-        const updatedUser = await userManager.update(uid, data);
+        const updatedProduct = await productManager.update(pid, data);
         return res.json({
             statusCode: 200,
-            message: `User with id ${uid} updated successfully`,
-            response: updatedUser
+            message: `Product with id ${pid} updated successfully`,
+            response: updatedProduct
         });    
     } catch(err) {
         return next(err);
@@ -83,16 +84,17 @@ async function update (req, res, next) {
 
 async function destroy (req, res, next) {
     try {
-        const { uid } = req.params;
-        const deletedUser = await userManager.destroy(uid);
+        const { pid } = req.params;
+        const deletedProduct = await productManager.destroy(pid);
         return res.json({
             statusCode: 200,
-            message: 'User deleted successfully',
-            response: deletedUser
+            message: 'Product deleted successfully',
+            response: deletedProduct
         });
     } catch(err) {
         return next(err);
     };
 };
 
-export default usersRouter;
+export default productsRouter;
+
