+++ src/views/index.handlebars
@@ -1,0 +1,7 @@
+<h1 class="bg-success-subtle p-2 text-center">HOME</h1>
+
+{{#each allProducts}}
+    <a href="/products/detail/{{this.id}}" class="pt-3 ps-3">{{this.title}}</a>
+{{/each}}
+
+{{!-- Falta Logo --}}
