+++ src/views/productDetail.handlebars
@@ -1,0 +1,7 @@
+{{#if product}}
+    <h3>About {{product.title}}</h3>
+    <p>Price: {{product.price}}</p>
+    <p>category: {{product.category}}</p>
+{{else}}
+    <h1>Product not found</h1>
+{{/if}}
