+++ src/views/register.handlebars
@@ -1,0 +1,11 @@
+{{!-- form de registro solo estructura --}}
+<h1>{{title}}</h1>
+<article class="w-50 d-flex flex-column align-items-center">
+    <h2 class="m-3">Register</h2>
+    <form class="d-flex flex-column">
+        <input class="m-3 w-40" type="email" name="email" id="email" placeholder="Email">
+        <input class="m-3 w-40" type="password" name="password" id="password" placeholder="Password">       
+        <input class="m-3 w-40"  type="number" name="role" id="role" placeholder="Role">
+        <input class="m-3 w-80"  type="button" value="Submit" id="submitProduct">
+    </form>
+</article>
